total price added

+++ src/components/AppHeader.tsx
@@ -1,6 +1,5 @@
 "use client";
 
-import { useEffect, useState } from "react";
 import useLocalStorage from "@/config/hooks/useLocalStorage.hooks";
 import { useAppDispatch } from "@/store";
 import { toggleLoginDialogState } from "@/store/slices/auth.slice";
@@ -8,7 +7,7 @@ import { appLogo } from "@assets/images/home";
 import { LocalStorageKeys } from "@lib/constants";
 import Image from "next/image";
 import Link from "next/link";
-import { root_container } from "@/app/Providers";
+import { useEffect, useLayoutEffect, useState } from "react";
 import { AppNavBar } from "./AppNavBar";
 import AppSidebar from "./AppSidebar";
 
@@ -16,13 +15,17 @@ const Header = () => {
   const dispatch = useAppDispatch();
   const [accessToken, __, removeAccessToken] = useLocalStorage(LocalStorageKeys.ACCESS_TOKEN, "");
   const [isClient, setIsClient] = useState(false);
+  const [buttonLable, setButtonLable] = useState("logout");
 
-  useEffect(() => {
+  useLayoutEffect(() => {
     setIsClient(true);
   }, []);
+  useEffect(() => {
+    accessToken ? setButtonLable("logout") : setButtonLable("login");
+  }, [accessToken]);
 
   const handleLoginLogout = () => {
-    if (accessToken && typeof window !== "undefined") {
+    if (isClient && accessToken && typeof window !== "undefined") {
       removeAccessToken();
       localStorage.clear();
     } else {
@@ -48,7 +51,7 @@ const Header = () => {
               onClick={handleLoginLogout}
               className="bg-primary/70 text-white py-2 px-6 rounded-lg flex items-center justify-center shadow-md hover:bg-primary transition duration-300"
             >
-              {accessToken ? "Logout" : "Login"}
+              {buttonLable}
             </button>
           )}
         </div>
